Fix start_tcplistener

+++ o2versioner/src/sequencer/handler.rs
@@ -12,7 +12,7 @@ use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
 type ArcState = Arc<Mutex<State>>;
 
 /// Main entrance for the sequencer
-pub async fn main<A>(addr: A, max_connection: Option<usize>)
+pub async fn main<A>(addr: A, max_connection: Option<u32>)
 where
     A: ToSocketAddrs + std::fmt::Debug + Clone,
 {
